==> python/LLP Accounts 2022/0.0 Read This First.ipynb <==
Overview of process for generating TaxCalc journals for LLP accounts

0.1 LLPAccounts_StitchTogetherBankAndBeals
- this uses Rent Statement spreadsheet to create spreadsheets 6045.xls, 3072.xls, beals.xls
- rent statments are used because property codess are already allocated - in future when not using property app, you will just need to create a bank download spreadsheet for each month and manually allocate property codes (though might be possible to do automatically).
    
1.0 LLPAccounts_AllocatePropertyCodesAndCategories
- this allocates property codes to Beals rent payments and mortgages
- generates 6045_categorised.xlsx and 3072_categorised.xlsx
- **** must check categories are correct and that all transactions have categories ****
    
2.0 LLPAccounts_ReconcileBankWithBeals

3.0 LLPAccounts_GenerateJournals

4.0 LLPAccounts_CompareBankDataToPropertyAppExport
    
this is to check the validity of data in property-application
- correct subcategories of expenses using sql scripts in LLP Accounts folder
    - GSC, Legal & Compliance, 
- add Kingson Rd mortgage (if it exists)
- add Birmingham Midshires small mortgages on F2 321 and
