==> python/LLP Accounts 2022/2.0 LLPAccounts_ReconcileBankWithBeals.ipynb <==
### Reconcile Bank data with Beals data

- Check Beals cumulative payments
    - puts credits/debits into a single column and keeps track of cumulative values - a records with an 'Amount payable' 
      with a non-zero value shows records have been carried forward from previous year or are to be carried forward to
      next.
- Reconcile Bank payments against Beals
    - iterate bank data for each payment from Beals, checking for corresponding data in Beals transactions list.
    - check for Bank records not reconciled to Beals payment
    - check for Beals records with not corresponding Bank payment
    - find Beals payments that include a transaction from previous year (identified by non zero cumulative amount)
    - find Beals transactions that have not been paid for and so are carried forward to next year (reconcile=0 flag)
- Create output spreadsheets
    - included is combo ledger of bank payments received along with Beals transactions in order to check balances

import numpy as np
import pandas as pd

import time
import os.path
import shutil

def backup_file(filename):
    timestr = time.strftime('%Y%m%d-%H%M%S')
    if os.path.isfile(filename):
        shutil.copy2(filename,filename+'.bak_' + timestr)
    return filename

#path = 'M:/My Documents/Business/Bugisiw Ltd/TaxReturn/TaxReturn_2020-2021/LLP accounts/'
taxreturn_yr='TaxReturn_2022-2023'
path = 'E:\\dtuklaptop\\e\\Users\\Mat\\python\\data\\property\\'+taxreturn_yr+'\\LLP accounts\\'
startdatestr='2022-04-06'
enddatestr='2022-07-31'

# Load bank files
#df6=pd.read_csv(path+'6045_s1.csv', parse_dates=['Date'], dayfirst=True)
#df3=pd.read_csv(path+'3072_s1.csv', parse_dates=['Date'], dayfirst=True)
input_file_6045 = path + '6045_categorised.xlsx'
df6=pd.read_excel(input_file_6045,sheet_name='6045')
df6=df6[~df6.Account.isnull()]
input_file_3072 = path + '3072_categorised.xlsx'
df3=pd.read_excel(input_file_3072,sheet_name='3072')
df3=df3[~df3.Account.isnull()]

# Load Beals transactions
#dfB=pd.read_csv(path+'beals.csv', parse_dates=['Date Created'], dayfirst=True)
dfB=pd.read_excel(path+'beals.xlsx', sheet_name='beals', parse_dates=['Date Created'], dayfirst=True)
dfB.insert(0,'Date', dfB['Date Created'].dt.date)
dfB=dfB.sort_values(by=['ID','Date Created'])
dfB['Date'] =  pd.to_datetime(dfB.Date)
#dfB['Item Description'] = dfB.ID + ': ' + dfB['Item Description']   

startdate = pd.to_datetime(startdatestr).date()
enddate = pd.to_datetime(enddatestr).date()
dfB=dfB[(dfB.Date >= startdate)&(dfB.Date <= enddate)]
df6=df6[(df6.Date >= startdate)&(df6.Date <= enddate)]
df3=df3[(df3.Date >= startdate)&(df3.Date <= enddate)]

### Add cumulative payments to Beals transactions
- it's easier to see missing transacations when there is a cumulative figure

# Create amount column
for index, row in dfB.iterrows():
    if(row['Credit Amount']>0):
        dfB.at[index,'Amount']=-round(row['Credit Amount'],2)
    elif(row['Debit Amount']>0):
        dfB.at[index,'Amount']=round(row['Debit Amount'],2)

# Track cumulative amounts between payments
cumamount=0.0;
lastID=''
for index, row in dfB.iterrows():
    if(row.ID != lastID):
        cumamount=row.Amount
    else:
        cumamount=cumamount + row.Amount

    dfB.at[index,'CumAmount'] = round(cumamount,2)
    if('Amount payable' in row['Item Description']):
        cumamount=0.0
    lastID=row.ID

### Reconcile Bank payments against Beals

Iterate bank data for each payment from Beals, checking for corresponding data in Beals transactions list.

Mark reconciled data in bank and reconciled Beals payments with 1, mark Beals transactions with account they were paid into.

from datetime import datetime, timedelta

def find_last_rent_received(payment_index, property_code, df_beals):
    rec=df_beals[(df_beals.ID==property_code)&(df_beals['Item Description'].str.contains('Rent for period'))&(df_beals.index<payment_index)]
    if not rec.empty:
        return rec['Item Description'].iloc[-1]
    else:
        return 'NOT FOUND'

def reconcile_bank_with_beals(df_bank, df_beals,bank_label):
    lookback = 6
    errorcount = 0
    for index, row in df_bank[df_bank['Memo'].str.contains("BEALS ESTATE AGENT", na=False)].iterrows():
        # Lookback 5 days for payment to appear in Beals statement
        print('Reconciling ' + str(row.Date) + ' ' + row.Memo + ' ' + str(row.Amount))

        startdate = row.Date - timedelta(days=lookback)
        enddate = row.Date
        rec=df_beals.loc[(df_beals['ID']==row.Property)&(df_beals['Debit Amount']==row.Amount)&(df_beals['Date']>=startdate)&(df_beals['Date']<=enddate)&(df_beals.ReconciledBank!=1)]
        if not rec.empty:
            print( str(rec.iloc[0].name)+' With '+ str(rec.Date.iloc[0]) + ' ' + rec['Item Description'].iloc[0] + ' ' + str(rec['Debit Amount'].iloc[0]))
            df_bank.at[index, 'ReconciledBeals']=1
            df_bank.at[index, 'Description']=find_last_rent_received(rec.iloc[0].name,row.Property,df_beals)
            df_beals.at[rec.iloc[0].name, 'ReconciledBank']=1
            df_beals.at[rec.iloc[0].name, 'ReconciledBankAccount']=bank_label
        else:
            print('CANNOT RECONCILE')
            errorcount = errorcount + 1

    print('Errors=' +  str(errorcount))

df6.loc[df6['Memo'].str.contains("BEALS ESTATE AGENT", na=False),'ReconciledBeals']=0
df3.loc[df3['Memo'].str.contains("BEALS ESTATE AGENT", na=False),'ReconciledBeals']=0
dfB['ReconciledBank']=np.nan
dfB['ReconciledBankAccount']=np.nan

#### Reconcile 6045 with Beals

reconcile_bank_with_beals(df6,dfB,6045)

#### Reconcile 3072 with Beals

reconcile_bank_with_beals(df3,dfB,3072)

#### Tidy reconciliation flag

# Backfill reconciliation flag so we know what bank account all Beals transactions relate to
dfB['ReconciledBank'] = dfB.groupby('ID')['ReconciledBank'].bfill()
dfB['ReconciledBank'].fillna(value=0, inplace=True)
dfB['ReconciledBankAccount'] = dfB.groupby('ID')['ReconciledBankAccount'].bfill()
dfB['ReconciledBankAccount'] = dfB.groupby('ID')['ReconciledBankAccount'].ffill()

#### Error check - Bank transactions not reconciled to Beals Payments

- check Beals data is complete if there are bank payments without a Beals record

df6.loc[(df6['Memo'].str.contains("BEALS ESTATE AGENT"))&(df6.ReconciledBeals==0)]

df3.loc[(df3['Memo'].str.contains("BEALS ESTATE AGENT"))&(df3.ReconciledBeals==0)]

#### Error check - Beals Payments not reconciled to Bank transaction - BEALS SAID THEY PAID BUT NOTHING RECEIVED
- there could be Beals payments at end of year which were paid into bank after year end
- need to contact Beals if there payments on other dates 

dfB.loc[(dfB['Item Description'].str.contains("Amount payable"))&(dfB['Item Description'].str.contains("Tucker"))&(dfB.ReconciledBank==0)]

dfB.loc[(dfB['Item Description'].str.contains("Amount payable"))&(dfB['Item Description'].str.contains("Valentino"))&(dfB.ReconciledBank==0)]

#### Payments for Beals transactions from previous year

This section works on the idea that when an amount is payed from Beals ("Amount payable"), this payment should clear the cumulative amount back to 0. If it is not 0, then there is something that might need looking at.

The dates on below transactions should all be around start of period. The cumulative amounts of £6 indicate Legislation & Compliance charges in previous period but paid for in this period. Dates later in year can indicate missing data in Beals spreadsheet so check the Beals paper statements match the Beals data.

dfB.loc[(dfB['Item Description'].str.contains("Amount payable"))&(dfB.CumAmount!=0)&(dfB.ReconciledBankAccount==6045)].sort_values(by=['Date'])

**** It looks like Flat 10 12-14 went wrong somewhere ****
- I checked this and the problem is things are a bit in the wrong order in Beals transactions files - it balances out to 0 in the end

dfB['Amount'].loc[(dfB['Item Description'].str.contains("Amount payable"))&(dfB.CumAmount!=0)&(dfB.ReconciledBankAccount==6045)].sum()

# Output to file
#dfB.loc[(dfB['Item Description'].str.contains("Amount payable"))&(dfB.CumAmount!=0)&(dfB.ReconciledBankAccount==6045)].sort_values(by=['Date']).to_excel(path+'6045_prevyr.xlsx',sheet_name='6045', index=False)

dfB.loc[(dfB['Item Description'].str.contains("Amount payable"))&(dfB.CumAmount!=0)&(dfB.ReconciledBankAccount==3072)].sort_values(by=['Date'])

dfB['Amount'].loc[(dfB['Item Description'].str.contains("Amount payable"))&(dfB.CumAmount!=0)&(dfB.ReconciledBankAccount==3072)].sum()

# Output to file
#dfB.loc[(dfB['Item Description'].str.contains("Amount payable"))&(dfB.CumAmount!=0)&(dfB.ReconciledBankAccount==3072)].sort_values(by=['Date']).to_excel(path+'3072_prevyr.xlsx',sheet_name='3072', index=False)

#### Payments carried forward to next year
- Often the records in the middle of the year are a result of Beals charging Legal & Compliance fees for flats which we took over after the tenant moved out

dfB.loc[(dfB.ReconciledBank==0)&(dfB.ReconciledBankAccount==6045)].sort_values(by=['ID','Date'])

dfB['Amount'].loc[(dfB.ReconciledBank==0)&(dfB.ReconciledBankAccount==6045)].sum()

# Output carried forward to file
dfB.loc[(dfB.ReconciledBank==0)&(dfB.ReconciledBankAccount==6045)].sort_values(by=['Date']).to_excel(path+'6045_carriedfwd.xlsx',sheet_name='6045', index=False)

dfB.loc[(dfB.ReconciledBank==0)&(dfB.ReconciledBankAccount==3072)].sort_values(by=['ID','Date'])

dfB['Amount'].loc[(dfB.ReconciledBank==0)&(dfB.ReconciledBankAccount==3072)].sum()

dfB.loc[(dfB.ReconciledBank==0)&(dfB.ReconciledBankAccount==3072)].sort_values(by=['Date']).to_excel(path+'3072_carriedfwd.xlsx',sheet_name='3072', index=False)

### Output files
- save Bank files with Beals payment reconciliation flag
- save Beals files with bank reconciliation flag, bank account it was paid into and cumulative amount

file6=backup_file(path+'6045_categorisedAndBealsReconciled.xlsx')
file3=backup_file(path+'3072_categorisedAndBealsReconciled.xlsx')
fileB=backup_file(path+'beals_bankReconciled.xlsx')
df6.to_excel(file6,sheet_name='6045', index=False)
df3.to_excel(file3,sheet_name='3072', index=False)
dfB.to_excel(fileB,sheet_name='beals', index=False)

### Create Beals/Bank reconciliation check
- this file can be used to check that the bank deposits minus agency fees match the Beals payments

dfB.rename(columns={'ReconciledBankAccount':'BankAccount'},inplace=True)
dfB.rename(columns={'Item Description':'PostNarrative'},inplace=True)
dfB.rename(columns={'ID':'Property'},inplace=True)

dfB.loc[(dfB.PostNarrative.str.contains('Rent for period'))&(dfB.BankAccount==6045),'Account']='0001'
dfB.loc[(dfB.PostNarrative.str.contains('Rent for period'))&(dfB.BankAccount==6045),'Account Description']='Rental Income'
dfB.loc[(dfB.PostNarrative.str.contains('Rent for period'))&(dfB.BankAccount==6045),'Category']='RentalIncome'
dfB.loc[(dfB.PostNarrative.str.contains('Rent for period'))&(dfB.BankAccount==3072),'Account']='0002'
dfB.loc[(dfB.PostNarrative.str.contains('Rent for period'))&(dfB.BankAccount==3072),'Account Description']='Rental Income'
dfB.loc[(dfB.PostNarrative.str.contains('Rent for period'))&(dfB.BankAccount==3072),'Category']='RentalIncome'

dfB.loc[(dfB.PostNarrative.str.contains('Legislation & Compliance'))&(dfB.BankAccount==6045),'Account']='1570'
dfB.loc[(dfB.PostNarrative.str.contains('Legislation & Compliance'))&(dfB.BankAccount==6045),'Account Description']='Management Fees'
dfB.loc[(dfB.PostNarrative.str.contains('Legislation & Compliance'))&(dfB.BankAccount==6045),'Category']='AgentFees'
dfB.loc[(dfB.PostNarrative.str.contains('Legislation & Compliance'))&(dfB.BankAccount==3072),'Account']='1570'
dfB.loc[(dfB.PostNarrative.str.contains('Legislation & Compliance'))&(dfB.BankAccount==3072),'Account Description']='Management Fees'
dfB.loc[(dfB.PostNarrative.str.contains('Legislation & Compliance'))&(dfB.BankAccount==3072),'Category']='AgentFees'

dfB.loc[(dfB.PostNarrative.str.contains('Management Fee'))&(dfB.BankAccount==6045),'Account']='1570'
dfB.loc[(dfB.PostNarrative.str.contains('Management Fee'))&(dfB.BankAccount==6045),'Account Description']='Management Fees'
dfB.loc[(dfB.PostNarrative.str.contains('Management Fee'))&(dfB.BankAccount==6045),'Category']='AgentFees'
dfB.loc[(dfB.PostNarrative.str.contains('Management Fee'))&(dfB.BankAccount==3072),'Account']='1570'
dfB.loc[(dfB.PostNarrative.str.contains('Management Fee'))&(dfB.BankAccount==3072),'Account Description']='Management Fees'
dfB.loc[(dfB.PostNarrative.str.contains('Management Fee'))&(dfB.BankAccount==3072),'Category']='AgentFees'

dfB.loc[(dfB.PostNarrative.str.contains('Amount payable'))&(dfB.BankAccount==6045),'Account']='4801'
dfB.loc[(dfB.PostNarrative.str.contains('Amount payable'))&(dfB.BankAccount==6045),'Account Description']='Payment'
dfB.loc[(dfB.PostNarrative.str.contains('Amount payable'))&(dfB.BankAccount==6045),'Category']='Payment'
dfB.loc[(dfB.PostNarrative.str.contains('Amount payable'))&(dfB.BankAccount==3072),'Account']='4801'
dfB.loc[(dfB.PostNarrative.str.contains('Amount payable'))&(dfB.BankAccount==3072),'Account Description']='Payment'
dfB.loc[(dfB.PostNarrative.str.contains('Amount payable'))&(dfB.BankAccount==3072),'Category']='Payment'

dfB.loc[(dfB.PostNarrative.str.contains('Amount payable'))&(dfB.BankAccount==6045),'Account']='4801'
dfB.loc[(dfB.PostNarrative.str.contains('Amount payable'))&(dfB.BankAccount==6045),'Account Description']='Payment'
dfB.loc[(dfB.PostNarrative.str.contains('Amount payable'))&(dfB.BankAccount==6045),'Category']='Payment'
dfB.loc[(dfB.PostNarrative.str.contains('Amount payable'))&(dfB.BankAccount==3072),'Account']='4801'
dfB.loc[(dfB.PostNarrative.str.contains('Amount payable'))&(dfB.BankAccount==3072),'Account Description']='Payment'
dfB.loc[(dfB.PostNarrative.str.contains('Amount payable'))&(dfB.BankAccount==3072),'Category']='Payment'

dfB.loc[(~dfB.PostNarrative.str.contains('Amount payable'))&(~dfB.PostNarrative.str.contains('Rent for period'))&(~dfB.PostNarrative.str.contains('Legislation & Compliance'))&(~dfB.PostNarrative.str.contains('Management Fee'))&(dfB.BankAccount==6045),'Account']='3140'
dfB.loc[(~dfB.PostNarrative.str.contains('Amount payable'))&(~dfB.PostNarrative.str.contains('Rent for period'))&(~dfB.PostNarrative.str.contains('Legislation & Compliance'))&(~dfB.PostNarrative.str.contains('Management Fee'))&(dfB.BankAccount==6045),'Account Description']='Repairs, renewals, maintenance'
dfB.loc[(~dfB.PostNarrative.str.contains('Amount payable'))&(~dfB.PostNarrative.str.contains('Rent for period'))&(~dfB.PostNarrative.str.contains('Legislation & Compliance'))&(~dfB.PostNarrative.str.contains('Management Fee'))&(dfB.BankAccount==6045),'Category']='Maintenance'
dfB.loc[(~dfB.PostNarrative.str.contains('Amount payable'))&(~dfB.PostNarrative.str.contains('Rent for period'))&(~dfB.PostNarrative.str.contains('Legislation & Compliance'))&(~dfB.PostNarrative.str.contains('Management Fee'))&(dfB.BankAccount==3072),'Account']='3140'
dfB.loc[(~dfB.PostNarrative.str.contains('Amount payable'))&(~dfB.PostNarrative.str.contains('Rent for period'))&(~dfB.PostNarrative.str.contains('Legislation & Compliance'))&(~dfB.PostNarrative.str.contains('Management Fee'))&(dfB.BankAccount==3072),'Account Description']='Repairs, renewals, maintenance'
dfB.loc[(~dfB.PostNarrative.str.contains('Amount payable'))&(~dfB.PostNarrative.str.contains('Rent for period'))&(~dfB.PostNarrative.str.contains('Legislation & Compliance'))&(~dfB.PostNarrative.str.contains('Management Fee'))&(dfB.BankAccount==3072),'Category']='Maintenance'


# Check nothing in Beals remains uncategorised
dfB[dfB.Category.isnull()]

cols=['BankAccount','Date','Property','PostNarrative','Amount','Account','Account Description','Category']

dfGr6=dfB.loc[(dfB.PostNarrative.str.contains('Rent for period'))&(dfB.BankAccount==6045)][cols]
dfGr3=dfB.loc[(dfB.PostNarrative.str.contains('Rent for period'))&(dfB.BankAccount==3072)][cols]
dfLc6=dfB.loc[(dfB.PostNarrative.str.contains('Legislation & Compliance'))&(dfB.BankAccount==6045)][cols]
dfLc3=dfB.loc[(dfB.PostNarrative.str.contains('Legislation & Compliance'))&(dfB.BankAccount==3072)][cols]
dfMg6=dfB.loc[(dfB.PostNarrative.str.contains('Management Fee'))&(dfB.BankAccount==6045)][cols]
dfMg3=dfB.loc[(dfB.PostNarrative.str.contains('Management Fee'))&(dfB.BankAccount==3072)][cols]
dfOth6=dfB.loc[(~dfB.PostNarrative.str.contains('Amount payable'))&(~dfB.PostNarrative.str.contains('Rent for period'))&(~dfB.PostNarrative.str.contains('Legislation & Compliance'))&(~dfB.PostNarrative.str.contains('Management Fee'))&(dfB.BankAccount==6045)][cols]
dfOth3=dfB.loc[(~dfB.PostNarrative.str.contains('Amount payable'))&(~dfB.PostNarrative.str.contains('Rent for period'))&(~dfB.PostNarrative.str.contains('Legislation & Compliance'))&(~dfB.PostNarrative.str.contains('Management Fee'))&(dfB.BankAccount==3072)][cols]

df6.rename(columns={'Account':'BankAccount'},inplace=True)
df6.rename(columns={'Memo':'PostNarrative'},inplace=True)
df6['BankAccount']=6045
df6['Account Description']='Payment'
df6.loc[(df6.Category=='BealsRentalIncome'),'Account']='4801'

df3.rename(columns={'Account':'BankAccount'},inplace=True)
df3.rename(columns={'Memo':'PostNarrative'},inplace=True)
df3['BankAccount']=3072
df3['Account Description']='Payment'
df3.loc[(df6.Category=='BealsRentalIncome'),'Account']='4801'

dfBk6=df6.loc[(df6.Category=='BealsRentalIncome')][cols]
dfBk3=df3.loc[(df3.Category=='BealsRentalIncome')][cols]

dfAll6=pd.concat([dfGr6,dfLc6,dfMg6,dfBk6,dfOth6]).sort_values(by=['Property','Date'])
dfAll6.reset_index(drop=True,inplace=True)
dfAll3=pd.concat([dfGr3,dfLc3,dfMg3,dfBk3,dfOth3]).sort_values(by=['Property','Date'])
dfAll3.reset_index(drop=True,inplace=True)

journals=path + 'journals\\'

beals_output_file=backup_file(journals + '3.6.AgentFeesRentalIncomeBankings_Beals_6045.xlsx')
writer = pd.ExcelWriter(beals_output_file, engine='xlsxwriter',datetime_format='dd/mm/yyyy')
dfGr6.to_excel(writer,sheet_name='GrossRent', index=False)
dfLc6.to_excel(writer,sheet_name='Legal', index=False)
dfMg6.to_excel(writer,sheet_name='MgtFee', index=False)
dfOth6.to_excel(writer,sheet_name='Other', index=False)
dfBk6.to_excel(writer,sheet_name='Bank', index=False)
dfAll6.to_excel(writer,sheet_name='All', index=False)
writer.save()

beals_output_file=backup_file(journals + '3.6.AgentFeesRentalIncomeBankings_Beals_3072.xlsx')
writer = pd.ExcelWriter(beals_output_file, engine='xlsxwriter',datetime_format='dd/mm/yyyy')
dfGr3.to_excel(writer,sheet_name='GrossRent', index=False)
dfLc3.to_excel(writer,sheet_name='Legal', index=False)
dfMg3.to_excel(writer,sheet_name='MgtFee', index=False)
dfOth3.to_excel(writer,sheet_name='Other', index=False)
dfBk3.to_excel(writer,sheet_name='Bank', index=False)
dfAll3.to_excel(writer,sheet_name='All', index=False)
writer.save()
