==> python/LLP Accounts 2022/3.0 LLPAccounts_GenerateJournals.ipynb <==
import numpy as np
import pandas as pd

import time
import os.path
import shutil

def backup_file(filename):
    timestr = time.strftime('%Y%m%d-%H%M%S')
    if os.path.isfile(filename):
        shutil.copy2(filename,filename+'.bak_' + timestr)
    return filename

#path = 'M:/My Documents/Business/Bugisiw Ltd/TaxReturn/TaxReturn_2020-2021/LLP accounts/'
taxreturn_yr='TaxReturn_2022-2023'
path = 'E:\\dtuklaptop\\e\\Users\\Mat\\python\\data\\property\\'+taxreturn_yr+'\\LLP accounts\\'
startdatestr='2022-04-06'
enddatestr='2022-07-31'

startdate = pd.to_datetime(startdatestr).date()
enddate = pd.to_datetime(enddatestr).date()

input_file_6045 = path + '6045_categorisedAndBealsReconciled.xlsx'
df6=pd.read_excel(input_file_6045,sheet_name='6045')
df6=df6[~df6.Account.isnull()]
input_file_3072 = path + '3072_categorisedAndBealsReconciled.xlsx'
df3=pd.read_excel(input_file_3072,sheet_name='3072')
df3=df3[~df3.Account.isnull()]
#df3.loc[(df3.Description.isnull()),'Description']=''

input_file_beals = path + 'beals_bankReconciled.xlsx'
dfB=pd.read_excel(input_file_beals,sheet_name='beals')
dfB['ReconciledBankAccount']=dfB['ReconciledBankAccount'].astype('str')

### Transform Bank data

df6.rename(columns={'Account':'BankAccount'},inplace=True)
df3.rename(columns={'Account':'BankAccount'},inplace=True)
df6['PostNarrative'] = df6.Property.replace(np.nan,'') + ':6045:' + df6['Memo'].replace(np.nan,'') + '/' + df6['Description'].replace(np.nan,'')
df3['PostNarrative'] = df3.Property.replace(np.nan,'') + ':3072:' + df3['Memo'].replace(np.nan,'') + '/' + df3['Description'].replace(np.nan,'')
df6['Amount']=-df6['Amount']
df3['Amount']=-df3['Amount']

### Assign Accounting codes

config_file = path + 'accounts-config.xlsx'
dfCfg=pd.read_excel(config_file,sheet_name='config',converters={'MT6045':str,'IV3072':str})

dfCfg

cols=['BankAccount','Date','Property','PostNarrative','Amount','Account','Account Description','Category']
journals=path + 'journals/'

for idx, row in dfCfg.iterrows():
    df6.loc[(df6.Category==row.Category),'Account']=row.MT6045
    df6.loc[(df6.Category==row.Category),'Account Description']=row.AccountDescription
    df3.loc[(df3.Category==row.Category),'Account']=row.IV3072
    df3.loc[(df3.Category==row.Category),'Account Description']=row.AccountDescription
    
    if(df6[cols][(df6.Category==row.Category)].shape[0]>0):
        writer = pd.ExcelWriter(backup_file(journals + '/' + str(row.Id) + '.' + row.Category + '_6045.xlsx'), engine='xlsxwriter',datetime_format='dd/mm/yyyy')
        df6[cols][(df6.Category==row.Category)].to_excel(writer,index=False)
        writer.save()
    if(df3[cols][(df3.Category==row.Category)].shape[0]>0):
        writer = pd.ExcelWriter(backup_file(journals + '/' + str(row.Id) + '.' + row.Category + '_3072.xlsx'), engine='xlsxwriter',datetime_format='dd/mm/yyyy')
        df3[cols][(df3.Category==row.Category)].to_excel(writer,index=False)
        writer.save()
