==> python/LLP Accounts 2022/4.0 LLPAccounts_CompareBankDataToPropertyAppExport.ipynb <==
import numpy as np
import pandas as pd

#path = 'M:/My Documents/Business/Bugisiw Ltd/TaxReturn/TaxReturn_2020-2021/LLP accounts/'
taxreturn_yr='TaxReturn_2022-2023'
path = 'E:\\dtuklaptop\\e\\Users\\Mat\\python\\data\\property\\'+taxreturn_yr+'\\LLP accounts\\'
startdatestr='2022-04-06'
enddatestr='2022-07-31'

propertyidmap = {
2:'321LON',
1:'F1321LON',
14:'F2321LON',
3:'F3321LON',
4:'F4321LON',
7:'169FAW',
5:'F1169FAW',
6:'F2169FAW',
8:'F3169FAW',
73:'F1171FAW',
74:'F2171FAW',
75:'F3171FAW',
12:'163FRA',
9:'F2163FRA',
10:'F3163FRA',
11:'F4163FRA',
17:'SHOP196KIN',
15:'196AKIN',
44:'196AKIN',
16:'196BKIN',
19:'23BHAM',
34:'23CHAM',
18:'23CHAM',
31:'F58ALH',
30:'F68ALH',
33:'F78ALH',
32:'F88ALH',
26:'F1746ALH',
22:'F1846ALH',
21:'F1946ALH',
25:'F2046ALH',
28:'F2146ALH',
23:'F2246ALH',
27:'F2346ALH',
46:'F31214ALH',
52:'F41214ALH',
47:'F51214ALH',
48:'F71214ALH',
49:'F111214ALH',
53:'F101214ALH',
54:'F141214ALH',
55:'F161214ALH',
66:'F11618ALH',
67:'F31618ALH',
63:'F61618ALH',
68:'F71618ALH',
64:'F111618ALH',
71:'F121618ALH',
70:'F131618ALH',
65:'F161618ALH',
81:'RSA'
}

startdate = pd.to_datetime(startdatestr).date()
enddate = pd.to_datetime(enddatestr).date()

input_file_6045 = path + '6045_categorisedAndBealsReconciled.xlsx'
df6=pd.read_excel(input_file_6045,sheet_name='6045')
df6=df6[~df6.Account.isnull()]
df6=df6[(df6.Date >= startdate)&(df6.Date <= enddate)]

input_file_3072 = path + '3072_categorisedAndBealsReconciled.xlsx'
df3=pd.read_excel(input_file_3072,sheet_name='3072')
df3=df3[~df3.Account.isnull()]
df3=df3[(df3.Date >= startdate)&(df3.Date <= enddate)]

The following file is exported from Property App view INCOME_EXPENDITURE_TAX_LLP_ALL for this tax year.

Don't forget to add the header row.

input_file_beals = path + 'income_expenditure_tax_llp_all_3.xls'
dfR=pd.read_excel(input_file_beals,sheet_name='Sheet 1')
dfR=dfR[(dfR.date >= startdate)&(dfR.date <= enddate)]

for index, row in dfR.iterrows():
    dfR.at[index, 'Property']=propertyidmap[row.property_id]

### Check Bank data against Rent Statements
- differences in values indicate that some items may be included in Rent Statements that are missing from Bank
- INVESTIGATE EACH PROPERTY WHERE THE DIFFERENCE IS NOT ZERO - this is a data integrity check

dfBank=pd.concat([df6,df3]).sort_values(by=['Date'])
dfBG=dfBank.groupby('Property')['Amount'].sum()

#### Check all properties - view difference between property-application and bank downloads

dfRG=dfR.groupby('Property')['amount'].sum()

dfCheck=pd.concat([dfBG, dfRG], axis=1)
dfCheck.rename(columns={'Amount':'AmountB'},inplace=True)
dfCheck.rename(columns={'amount':'AmountRS'},inplace=True)
dfCheck['Diff']=round(dfCheck.AmountB - dfCheck.AmountRS,2)
dfCheck

196 GFF Rent registered in papp as 05/04/22 (before bank data) but received in bank next day - so ignore error as it crosses period.

F10 12-14 - I think Beals data is wrong. This flat had errors with cumulative payments in other notebook. I've checked bank and four months rent have be paid. Rent statement is saying 2 more months have been paid which is probably not right unless he was in arrears. Will as for bank download for this flat to check this.

F6 8 - £47.57 in papp was paid after bank period end (31/07/22), also £6 carried fwd from prev period

F7 8 - this is an overlap of period like for 196 GFF.

#### Check individual properties (to pin point problems)

pty='F68ALH'

dfBank.loc[dfBank.Property==pty].groupby(['Category'])['Amount'].sum()

dfR.loc[(dfR.Property==pty)].groupby(['cat'])['amount'].sum()

dfR[(dfR.Property==pty)&(dfR.cat.str.contains('Lettings Agency|Tenancy Related'))].loc[dfR.Property==pty]['amount'].sum()

dfR[(dfR.Property==pty)&(dfR.cat.str.contains('Maintenance|Utilities'))].loc[dfR.Property==pty]['amount'].sum()

### Overall view of bank data

Total figures for both accounts followed by individual accounts.

dfBank.groupby(['Category'])['Amount'].sum()

df6.groupby(['Category'])['Amount'].sum()

df3.groupby(['Category'])['Amount'].sum()

Check monthly P/L:

input_file_6045 = path + '6045_categorisedAndBealsReconciled.xlsx'
df6=pd.read_excel(input_file_6045,sheet_name='6045')
df6=df6[~df6.Account.isnull()]
df6=df6[(df6.Date >= startdate)&(df6.Date <= enddate)]

input_file_3072 = path + '3072_categorisedAndBealsReconciled.xlsx'
df3=pd.read_excel(input_file_3072,sheet_name='3072')
df3=df3[~df3.Account.isnull()]
df3=df3[(df3.Date >= startdate)&(df3.Date <= enddate)]

#input_file_IV = path + '300 mvt .xlsx'
#dfIV=pd.read_excel(input_file_IV,sheet_name='data (36)')
#dfIV=dfIV[~dfIV.Account.isnull()]
#dfIV=dfIV[(dfIV.Date >= startdate)&(dfIV.Date <= enddate)]

#df6['Amount'][(df6.Date>='06/01/2020')&(df6.Date<='06/30/2020')].groupby(by=[df6.Date.dt.date]).sum()
#dfIV['Amount'][(dfIV.Date>='06/01/2020')&(dfIV.Date<='06/30/2020')].groupby(by=[dfIV.Date.dt.date]).sum()

df6['Amount'].groupby(by=[df6.Date.dt.year, df6.Date.dt.month]).sum()

df3['Amount'].groupby(by=[df3.Date.dt.year, df3.Date.dt.month]).sum()

df6['Amount'].sum()

df3['Amount'].sum()
